+++ shroud/wrapl.py
@@ -457,10 +457,12 @@ luaL_setfuncs({LUA_state_var}, {LUA_class_reg}, 0);
             fmt_result0 = node._fmtresult
             fmt_result = fmt_result0.setdefault('fmtl', util.Scope(fmt))
             fmt_result.cxx_var = wformat('{CXX_local}{LUA_result}', fmt_result)
-            if ast.is_pointer():
+            if is_ctor or ast.is_pointer():
                 fmt_result.cxx_deref = '->'
+                fmt_result.cxx_addr = ''
             else:
                 fmt_result.cxx_deref = '.'
+                fmt_result.cxx_addr = '&'
             if result_typedef.cxx_to_c:
                 fmt_result.c_var = wformat(result_typedef.cxx_to_c, fmt_result)  # if C++
             else: