+++ tests/tutorial/pySingletontype.cpp
@@ -1,4 +1,4 @@
-// pyNames2type.cpp
+// pySingletontype.cpp
 // This is generated code, do not edit
 // #######################################################################
 // Copyright (c) 2017-2018, Lawrence Livermore National Security, LLC.
@@ -40,9 +40,10 @@
 // SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 //
 // #######################################################################
-#include "pytestnamesmodule.hpp"
-// splicer begin class.Names2.impl.include
-// splicer end class.Names2.impl.include
+#include "pyTutorialmodule.hpp"
+#include "tutorial.hpp"
+// splicer begin class.Singleton.impl.include
+// splicer end class.Singleton.impl.include
 
 #ifdef __cplusplus
 #define SHROUD_UNUSED(param)
@@ -55,35 +56,59 @@
 #define PyString_FromString PyUnicode_FromString
 #define PyString_FromStringAndSize PyUnicode_FromStringAndSize
 #endif
-// splicer begin class.Names2.impl.C_definition
-// splicer end class.Names2.impl.C_definition
-// splicer begin class.Names2.impl.additional_methods
-// splicer end class.Names2.impl.additional_methods
+// splicer begin class.Singleton.impl.C_definition
+// splicer end class.Singleton.impl.C_definition
+// splicer begin class.Singleton.impl.additional_methods
+// splicer end class.Singleton.impl.additional_methods
 static void
-PY_Names2_tp_del (PY_Names2 *self)
+PY_Singleton_tp_del (PY_Singleton *self)
 {
-// splicer begin class.Names2.type.del
+// splicer begin class.Singleton.type.del
     delete self->obj;
     self->obj = NULL;
-// splicer end class.Names2.type.del
+// splicer end class.Singleton.type.del
 }
-// splicer begin class.Names2.impl.after_methods
-// splicer end class.Names2.impl.after_methods
-static PyMethodDef PY_Names2_methods[] = {
-    // splicer begin class.Names2.PyMethodDef
-    // splicer end class.Names2.PyMethodDef
+
+static char PY_singleton_getReference__doc__[] =
+"documentation"
+;
+
+static PyObject *
+PY_singleton_getReference(
+  PY_Singleton *self,
+  PyObject *SHROUD_UNUSED(args),
+  PyObject *SHROUD_UNUSED(kwds))
+{
+// Singleton & getReference()
+// splicer begin class.Singleton.method.get_reference
+    Singleton & SHCXX_rv = Singleton::getReference();
+
+    // post_call
+    PY_Singleton * SHTPy_rv = PyObject_New(PY_Singleton, &PY_Singleton_Type);
+    SHTPy_rv->obj = &SHCXX_rv;
+
+    return (PyObject *) SHTPy_rv;
+// splicer end class.Singleton.method.get_reference
+}
+// splicer begin class.Singleton.impl.after_methods
+// splicer end class.Singleton.impl.after_methods
+static PyMethodDef PY_Singleton_methods[] = {
+    {"getReference", (PyCFunction)PY_singleton_getReference,
+        METH_STATIC|METH_NOARGS, PY_singleton_getReference__doc__},
+    // splicer begin class.Singleton.PyMethodDef
+    // splicer end class.Singleton.PyMethodDef
     {NULL,   (PyCFunction)NULL, 0, NULL}            /* sentinel */
 };
 
-static char Names2__doc__[] =
+static char Singleton__doc__[] =
 "virtual class"
 ;
 
 /* static */
-PyTypeObject PY_Names2_Type = {
+PyTypeObject PY_Singleton_Type = {
         PyVarObject_HEAD_INIT(NULL, 0)
-        "testnames.Names2",                       /* tp_name */
-        sizeof(PY_Names2),         /* tp_basicsize */
+        "tutorial.Singleton",                       /* tp_name */
+        sizeof(PY_Singleton),         /* tp_basicsize */
         0,                              /* tp_itemsize */
         /* Methods to implement standard operations */
         (destructor)0,                 /* tp_dealloc */
@@ -110,7 +135,7 @@ PyTypeObject PY_Names2_Type = {
         0,                              /* tp_as_buffer */
         /* Flags to define presence of optional/expanded features */
         Py_TPFLAGS_DEFAULT,             /* tp_flags */
-        Names2__doc__,         /* tp_doc */
+        Singleton__doc__,         /* tp_doc */
         /* Assigned meaning in release 2.0 */
         /* call function for all accessible objects */
         (traverseproc)0,                /* tp_traverse */
@@ -126,7 +151,7 @@ PyTypeObject PY_Names2_Type = {
         (getiterfunc)0,                 /* tp_iter */
         (iternextfunc)0,                /* tp_iternext */
         /* Attribute descriptor and subclassing stuff */
-        PY_Names2_methods,                             /* tp_methods */
+        PY_Singleton_methods,                             /* tp_methods */
         0,                              /* tp_members */
         0,                             /* tp_getset */
         0,                              /* tp_base */
@@ -144,7 +169,7 @@ PyTypeObject PY_Names2_Type = {
         0,                              /* tp_cache */
         0,                              /* tp_subclasses */
         0,                              /* tp_weaklist */
-        (destructor)PY_Names2_tp_del,                 /* tp_del */
+        (destructor)PY_Singleton_tp_del,                 /* tp_del */
         0,                              /* tp_version_tag */
 #if PY_MAJOR_VERSION >= 3
         (destructor)0,                  /* tp_finalize */