+++ tests/tutorial/pyTutorialmodule.hpp
@@ -50,16 +50,21 @@
 namespace tutorial {
     class Class1;
 }
+class Singleton;
 
 extern PyTypeObject PY_Class1_Type;
+extern PyTypeObject PY_Singleton_Type;
 
 // splicer begin header.C_declaration
 // splicer end header.C_declaration
 
 // helper functions
 extern const char *PY_Class1_capsule_name;
+extern const char *PY_Singleton_capsule_name;
 PyObject *PP_Class1_to_Object(tutorial::Class1 *addr);
 int PP_Class1_from_Object(PyObject *obj, void **addr);
+PyObject *PP_Singleton_to_Object(Singleton *addr);
+int PP_Singleton_from_Object(PyObject *obj, void **addr);
 
 // splicer begin class.Class1.C_declaration
 // splicer end class.Class1.C_declaration
@@ -70,6 +75,15 @@ PyObject_HEAD
     // splicer begin class.Class1.C_object
     // splicer end class.Class1.C_object
 } PY_Class1;
+// splicer begin class.Singleton.C_declaration
+// splicer end class.Singleton.C_declaration
+
+typedef struct {
+PyObject_HEAD
+    Singleton * obj;
+    // splicer begin class.Singleton.C_object
+    // splicer end class.Singleton.C_object
+} PY_Singleton;
 
 extern PyObject *PY_error_obj;