+++ tests/tutorial/wrapSingleton.cpp
@@ -1,4 +1,4 @@
-// pyTutorialhelper.cpp
+// wrapSingleton.cpp
 // This is generated code, do not edit
 // #######################################################################
 // Copyright (c) 2017-2018, Lawrence Livermore National Security, LLC.
@@ -40,65 +40,27 @@
 // SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 //
 // #######################################################################
-#include "pyTutorialmodule.hpp"
-const char *PY_Class1_capsule_name = "Class1";
-const char *PY_Singleton_capsule_name = "Singleton";
+#include "wrapSingleton.h"
+#include "tutorial.hpp"
 
+// splicer begin class.Singleton.CXX_definitions
+// splicer end class.Singleton.CXX_definitions
 
-PyObject *PP_Class1_to_Object(tutorial::Class1 *addr)
+extern "C" {
+
+// splicer begin class.Singleton.C_definitions
+// splicer end class.Singleton.C_definitions
+
+// Singleton & getReference()
+// function_index=5
+TUT_singleton * TUT_singleton_get_reference()
 {
-    // splicer begin class.Class1.helper.to_object
-    PyObject *voidobj;
-    PyObject *args;
-    PyObject *rv;
-
-    voidobj = PyCapsule_New(addr, PY_Class1_capsule_name, NULL);
-    args = PyTuple_New(1);
-    PyTuple_SET_ITEM(args, 0, voidobj);
-    rv = PyObject_Call((PyObject *) &PY_Class1_Type, args, NULL);
-    Py_DECREF(args);
-    return rv;
-    // splicer end class.Class1.helper.to_object
+// splicer begin class.Singleton.method.get_reference
+    Singleton & SHCXX_rv = Singleton::getReference();
+    TUT_singleton * SHC_rv = static_cast<TUT_singleton *>(
+        static_cast<void *>(&SHCXX_rv));
+    return SHC_rv;
+// splicer end class.Singleton.method.get_reference
 }
 
-int PP_Class1_from_Object(PyObject *obj, void **addr)
-{
-    // splicer begin class.Class1.helper.from_object
-    if (obj->ob_type != &PY_Class1_Type) {
-        // raise exception
-        return 0;
-    }
-    PY_Class1 * self = (PY_Class1 *) obj;
-    *addr = self->obj;
-    return 1;
-    // splicer end class.Class1.helper.from_object
-}
-
-PyObject *PP_Singleton_to_Object(Singleton *addr)
-{
-    // splicer begin class.Singleton.helper.to_object
-    PyObject *voidobj;
-    PyObject *args;
-    PyObject *rv;
-
-    voidobj = PyCapsule_New(addr, PY_Singleton_capsule_name, NULL);
-    args = PyTuple_New(1);
-    PyTuple_SET_ITEM(args, 0, voidobj);
-    rv = PyObject_Call((PyObject *) &PY_Singleton_Type, args, NULL);
-    Py_DECREF(args);
-    return rv;
-    // splicer end class.Singleton.helper.to_object
-}
-
-int PP_Singleton_from_Object(PyObject *obj, void **addr)
-{
-    // splicer begin class.Singleton.helper.from_object
-    if (obj->ob_type != &PY_Singleton_Type) {
-        // raise exception
-        return 0;
-    }
-    PY_Singleton * self = (PY_Singleton *) obj;
-    *addr = self->obj;
-    return 1;
-    // splicer end class.Singleton.helper.from_object
-}
+}  // extern "C"