+++ tests/tutorial/wrapTutorial.cpp
@@ -80,7 +80,7 @@ extern "C" {
 // splicer end C_definitions
 
 // void Function1()
-// function_index=5
+// function_index=6
 void TUT_function1()
 {
 // splicer begin function.function1
@@ -90,7 +90,7 @@ void TUT_function1()
 }
 
 // double Function2(double arg1 +intent(in)+value, int arg2 +intent(in)+value)
-// function_index=6
+// function_index=7
 double TUT_function2(double arg1, int arg2)
 {
 // splicer begin function.function2
@@ -100,7 +100,7 @@ double TUT_function2(double arg1, int arg2)
 }
 
 // void Sum(size_t len +implied(size(values))+intent(in)+value, int * values +dimension(:)+intent(in), int * result +intent(out))
-// function_index=7
+// function_index=8
 void TUT_sum(size_t len, int * values, int * result)
 {
 // splicer begin function.sum
@@ -110,7 +110,7 @@ void TUT_sum(size_t len, int * values, int * result)
 }
 
 // long long TypeLongLong(long long arg1 +intent(in)+value)
-// function_index=8
+// function_index=9
 long long TUT_type_long_long(long long arg1)
 {
 // splicer begin function.type_long_long
@@ -120,7 +120,7 @@ long long TUT_type_long_long(long long arg1)
 }
 
 // bool Function3(bool arg +intent(in)+value)
-// function_index=9
+// function_index=10
 bool TUT_function3(bool arg)
 {
 // splicer begin function.function3
@@ -130,7 +130,7 @@ bool TUT_function3(bool arg)
 }
 
 // void Function3b(const bool arg1 +intent(in)+value, bool * arg2 +intent(out), bool * arg3 +intent(inout))
-// function_index=10
+// function_index=11
 void TUT_function3b(const bool arg1, bool * arg2, bool * arg3)
 {
 // splicer begin function.function3b
@@ -140,7 +140,7 @@ void TUT_function3b(const bool arg1, bool * arg2, bool * arg3)
 }
 
 // void Function4a(const std::string & arg1 +intent(in)+len_trim(Larg1), const std::string & arg2 +intent(in)+len_trim(Larg2), std::string * SHF_rv +intent(out)+len(NSHF_rv)) +len(30)
-// function_index=47
+// function_index=48
 void TUT_function4a_bufferify(const char * arg1, int Larg1,
     const char * arg2, int Larg2, char * SHF_rv, int NSHF_rv)
 {
@@ -158,7 +158,7 @@ void TUT_function4a_bufferify(const char * arg1, int Larg1,
 }
 
 // const std::string & Function4b(const std::string & arg1 +intent(in), const std::string & arg2 +intent(in))
-// function_index=12
+// function_index=13
 const char * TUT_function4b(const char * arg1, const char * arg2)
 {
 // splicer begin function.function4b
@@ -172,7 +172,7 @@ const char * TUT_function4b(const char * arg1, const char * arg2)
 }
 
 // void Function4b(const std::string & arg1 +intent(in)+len_trim(Larg1), const std::string & arg2 +intent(in)+len_trim(Larg2), std::string & output +intent(out)+len(Noutput))
-// function_index=48
+// function_index=49
 void TUT_function4b_bufferify(const char * arg1, int Larg1,
     const char * arg2, int Larg2, char * output, int Noutput)
 {
@@ -191,7 +191,7 @@ void TUT_function4b_bufferify(const char * arg1, int Larg1,
 }
 
 // double Function5()
-// function_index=37
+// function_index=38
 double TUT_function5()
 {
 // splicer begin function.function5
@@ -201,7 +201,7 @@ double TUT_function5()
 }
 
 // double Function5(double arg1=3.1415 +intent(in)+value)
-// function_index=38
+// function_index=39
 double TUT_function5_arg1(double arg1)
 {
 // splicer begin function.function5_arg1
@@ -211,7 +211,7 @@ double TUT_function5_arg1(double arg1)
 }
 
 // double Function5(double arg1=3.1415 +intent(in)+value, bool arg2=true +intent(in)+value)
-// function_index=13
+// function_index=14
 double TUT_function5_arg1_arg2(double arg1, bool arg2)
 {
 // splicer begin function.function5_arg1_arg2
@@ -221,7 +221,7 @@ double TUT_function5_arg1_arg2(double arg1, bool arg2)
 }
 
 // void Function6(const std::string & name +intent(in))
-// function_index=14
+// function_index=15
 void TUT_function6_from_name(const char * name)
 {
 // splicer begin function.function6_from_name
@@ -232,7 +232,7 @@ void TUT_function6_from_name(const char * name)
 }
 
 // void Function6(const std::string & name +intent(in)+len_trim(Lname))
-// function_index=50
+// function_index=51
 void TUT_function6_from_name_bufferify(const char * name, int Lname)
 {
 // splicer begin function.function6_from_name_bufferify
@@ -243,7 +243,7 @@ void TUT_function6_from_name_bufferify(const char * name, int Lname)
 }
 
 // void Function6(int indx +intent(in)+value)
-// function_index=15
+// function_index=16
 void TUT_function6_from_index(int indx)
 {
 // splicer begin function.function6_from_index
@@ -253,7 +253,7 @@ void TUT_function6_from_index(int indx)
 }
 
 // void Function7(int arg +intent(in)+value)
-// function_index=39
+// function_index=40
 void TUT_function7_int(int arg)
 {
 // splicer begin function.function7_int
@@ -263,7 +263,7 @@ void TUT_function7_int(int arg)
 }
 
 // void Function7(double arg +intent(in)+value)
-// function_index=40
+// function_index=41
 void TUT_function7_double(double arg)
 {
 // splicer begin function.function7_double
@@ -273,7 +273,7 @@ void TUT_function7_double(double arg)
 }
 
 // int Function8()
-// function_index=41
+// function_index=42
 int TUT_function8_int()
 {
 // splicer begin function.function8_int
@@ -283,7 +283,7 @@ int TUT_function8_int()
 }
 
 // double Function8()
-// function_index=42
+// function_index=43
 double TUT_function8_double()
 {
 // splicer begin function.function8_double
@@ -293,7 +293,7 @@ double TUT_function8_double()
 }
 
 // void Function9(double arg +intent(in)+value)
-// function_index=18
+// function_index=19
 void TUT_function9(double arg)
 {
 // splicer begin function.function9
@@ -303,7 +303,7 @@ void TUT_function9(double arg)
 }
 
 // void Function10()
-// function_index=19
+// function_index=20
 void TUT_function10_0()
 {
 // splicer begin function.function10_0
@@ -313,7 +313,7 @@ void TUT_function10_0()
 }
 
 // void Function10(const std::string & name +intent(in), double arg2 +intent(in)+value)
-// function_index=20
+// function_index=21
 void TUT_function10_1(const char * name, double arg2)
 {
 // splicer begin function.function10_1
@@ -324,7 +324,7 @@ void TUT_function10_1(const char * name, double arg2)
 }
 
 // void Function10(const std::string & name +intent(in)+len_trim(Lname), double arg2 +intent(in)+value)
-// function_index=51
+// function_index=52
 void TUT_function10_1_bufferify(const char * name, int Lname,
     double arg2)
 {
@@ -336,7 +336,7 @@ void TUT_function10_1_bufferify(const char * name, int Lname,
 }
 
 // int overload1(int num +intent(in)+value)
-// function_index=43
+// function_index=44
 int TUT_overload1_num(int num)
 {
 // splicer begin function.overload1_num
@@ -346,7 +346,7 @@ int TUT_overload1_num(int num)
 }
 
 // int overload1(int num +intent(in)+value, int offset=0 +intent(in)+value)
-// function_index=44
+// function_index=45
 int TUT_overload1_num_offset(int num, int offset)
 {
 // splicer begin function.overload1_num_offset
@@ -356,7 +356,7 @@ int TUT_overload1_num_offset(int num, int offset)
 }
 
 // int overload1(int num +intent(in)+value, int offset=0 +intent(in)+value, int stride=1 +intent(in)+value)
-// function_index=21
+// function_index=22
 int TUT_overload1_num_offset_stride(int num, int offset, int stride)
 {
 // splicer begin function.overload1_num_offset_stride
@@ -366,7 +366,7 @@ int TUT_overload1_num_offset_stride(int num, int offset, int stride)
 }
 
 // int overload1(double type +intent(in)+value, int num +intent(in)+value)
-// function_index=45
+// function_index=46
 int TUT_overload1_3(double type, int num)
 {
 // splicer begin function.overload1_3
@@ -376,7 +376,7 @@ int TUT_overload1_3(double type, int num)
 }
 
 // int overload1(double type +intent(in)+value, int num +intent(in)+value, int offset=0 +intent(in)+value)
-// function_index=46
+// function_index=47
 int TUT_overload1_4(double type, int num, int offset)
 {
 // splicer begin function.overload1_4
@@ -386,7 +386,7 @@ int TUT_overload1_4(double type, int num, int offset)
 }
 
 // int overload1(double type +intent(in)+value, int num +intent(in)+value, int offset=0 +intent(in)+value, int stride=1 +intent(in)+value)
-// function_index=22
+// function_index=23
 int TUT_overload1_5(double type, int num, int offset, int stride)
 {
 // splicer begin function.overload1_5
@@ -396,7 +396,7 @@ int TUT_overload1_5(double type, int num, int offset, int stride)
 }
 
 // TypeID typefunc(TypeID arg +intent(in)+value)
-// function_index=23
+// function_index=24
 int TUT_typefunc(int arg)
 {
 // splicer begin function.typefunc
@@ -406,7 +406,7 @@ int TUT_typefunc(int arg)
 }
 
 // EnumTypeID enumfunc(EnumTypeID arg +intent(in)+value)
-// function_index=24
+// function_index=25
 int TUT_enumfunc(int arg)
 {
 // splicer begin function.enumfunc
@@ -418,7 +418,7 @@ int TUT_enumfunc(int arg)
 }
 
 // void getMinMax(int & min +intent(out), int & max +intent(out))
-// function_index=25
+// function_index=26
 /**
  * \brief Pass in reference to scalar
  *
@@ -432,7 +432,7 @@ void TUT_get_min_max(int * min, int * max)
 }
 
 // int useclass(const Class1 * arg1 +intent(in)+value)
-// function_index=26
+// function_index=27
 int TUT_useclass(const TUT_class1 * arg1)
 {
 // splicer begin function.useclass
@@ -444,7 +444,7 @@ int TUT_useclass(const TUT_class1 * arg1)
 }
 
 // const Class1 * getclass2()
-// function_index=27
+// function_index=28
 const TUT_class1 * TUT_getclass2()
 {
 // splicer begin function.getclass2
@@ -456,7 +456,7 @@ const TUT_class1 * TUT_getclass2()
 }
 
 // Class1 * getclass3()
-// function_index=28
+// function_index=29
 TUT_class1 * TUT_getclass3()
 {
 // splicer begin function.getclass3
@@ -468,7 +468,7 @@ TUT_class1 * TUT_getclass3()
 }
 
 // int vector_sum(const std::vector<int> & arg +dimension(:)+intent(in)+size(Sarg))
-// function_index=52
+// function_index=53
 int TUT_vector_sum_bufferify(const int * arg, long Sarg)
 {
 // splicer begin function.vector_sum_bufferify
@@ -479,7 +479,7 @@ int TUT_vector_sum_bufferify(const int * arg, long Sarg)
 }
 
 // void vector_iota(std::vector<int> & arg +dimension(:)+intent(out)+size(Sarg))
-// function_index=53
+// function_index=54
 void TUT_vector_iota_bufferify(int * arg, long Sarg)
 {
 // splicer begin function.vector_iota_bufferify
@@ -499,7 +499,7 @@ void TUT_vector_iota_bufferify(int * arg, long Sarg)
 }
 
 // void vector_increment(std::vector<int> & arg +dimension(:)+intent(inout)+size(Sarg))
-// function_index=54
+// function_index=55
 void TUT_vector_increment_bufferify(int * arg, long Sarg)
 {
 // splicer begin function.vector_increment_bufferify
@@ -519,7 +519,7 @@ void TUT_vector_increment_bufferify(int * arg, long Sarg)
 }
 
 // int vector_string_count(const std::vector<std::string> & arg +dimension(:)+intent(in)+len(Narg)+size(Sarg))
-// function_index=55
+// function_index=56
 /**
  * \brief count number of underscore in vector of strings
  *
@@ -545,7 +545,7 @@ int TUT_vector_string_count_bufferify(const char * arg, long Sarg,
 }
 
 // void vector_string_fill(std::vector<std::string> & arg +dimension(:)+intent(out)+len(Narg)+size(Sarg))
-// function_index=56
+// function_index=57
 /**
  * \brief Fill in arg with some animal names
  *
@@ -573,7 +573,7 @@ int TUT_vector_string_fill_bufferify(char * arg, long Sarg, int Narg)
 }
 
 // void vector_string_append(std::vector<std::string> & arg +dimension(:)+intent(inout)+len(Narg)+size(Sarg))
-// function_index=57
+// function_index=58
 /**
  * \brief append '-like' to names.
  *
@@ -609,7 +609,7 @@ void TUT_vector_string_append_bufferify(char * arg, long Sarg, int Narg)
 }
 
 // int callback1(int in +intent(in)+value, int ( * incr) +intent(in)+value(int +value))
-// function_index=35
+// function_index=36
 int TUT_callback1(int in, int ( * incr)(int))
 {
 // splicer begin function.callback1
@@ -619,7 +619,7 @@ int TUT_callback1(int in, int ( * incr)(int))
 }
 
 // const std::string & LastFunctionCalled() +len(30)
-// function_index=36
+// function_index=37
 const char * TUT_last_function_called()
 {
 // splicer begin function.last_function_called
@@ -630,7 +630,7 @@ const char * TUT_last_function_called()
 }
 
 // void LastFunctionCalled(std::string & SHF_rv +intent(out)+len(NSHF_rv)) +len(30)
-// function_index=58
+// function_index=59
 void TUT_last_function_called_bufferify(char * SHF_rv, int NSHF_rv)
 {
 // splicer begin function.last_function_called_bufferify