+++ tests/tutorial/wrapftutorial.f
@@ -73,12 +73,30 @@ module tutorial_mod
         ! splicer end class.Class1.type_bound_procedure_part
     end type class1
 
+    ! splicer begin class.Singleton.module_top
+    ! splicer end class.Singleton.module_top
+
+    type singleton
+        type(C_PTR), private :: voidptr
+        ! splicer begin class.Singleton.component_part
+        ! splicer end class.Singleton.component_part
+    contains
+        procedure, nopass :: get_reference => singleton_get_reference
+        procedure :: get_instance => singleton_get_instance
+        procedure :: set_instance => singleton_set_instance
+        procedure :: associated => singleton_associated
+        ! splicer begin class.Singleton.type_bound_procedure_part
+        ! splicer end class.Singleton.type_bound_procedure_part
+    end type singleton
+
     interface operator (.eq.)
         module procedure class1_eq
+        module procedure singleton_eq
     end interface
 
     interface operator (.ne.)
         module procedure class1_ne
+        module procedure singleton_ne
     end interface
 
     abstract interface
@@ -139,6 +157,17 @@ module tutorial_mod
 
         ! splicer begin class.Class1.additional_interfaces
         ! splicer end class.Class1.additional_interfaces
+
+        function c_singleton_get_reference() &
+                result(SHT_rv) &
+                bind(C, name="TUT_singleton_get_reference")
+            use iso_c_binding, only : C_PTR
+            implicit none
+            type(C_PTR) :: SHT_rv
+        end function c_singleton_get_reference
+
+        ! splicer begin class.Singleton.additional_interfaces
+        ! splicer end class.Singleton.additional_interfaces
 
         subroutine function1() &
                 bind(C, name="TUT_function1")
@@ -674,8 +703,42 @@ contains
     ! splicer begin class.Class1.additional_functions
     ! splicer end class.Class1.additional_functions
 
+    ! Singleton & getReference()
+    ! function_index=5
+    function singleton_get_reference() &
+            result(SHT_rv)
+        type(singleton) :: SHT_rv
+        ! splicer begin class.Singleton.method.get_reference
+        SHT_rv%voidptr = c_singleton_get_reference()
+        ! splicer end class.Singleton.method.get_reference
+    end function singleton_get_reference
+
+    function singleton_get_instance(obj) result (voidptr)
+        use iso_c_binding, only: C_PTR
+        class(singleton), intent(IN) :: obj
+        type(C_PTR) :: voidptr
+        voidptr = obj%voidptr
+    end function singleton_get_instance
+
+    subroutine singleton_set_instance(obj, voidptr)
+        use iso_c_binding, only: C_PTR
+        class(singleton), intent(INOUT) :: obj
+        type(C_PTR), intent(IN) :: voidptr
+        obj%voidptr = voidptr
+    end subroutine singleton_set_instance
+
+    function singleton_associated(obj) result (rv)
+        use iso_c_binding, only: c_associated
+        class(singleton), intent(IN) :: obj
+        logical rv
+        rv = c_associated(obj%voidptr)
+    end function singleton_associated
+
+    ! splicer begin class.Singleton.additional_functions
+    ! splicer end class.Singleton.additional_functions
+
     ! void Sum(size_t len +implied(size(values))+intent(in)+value, int * values +dimension(:)+intent(in), int * result +intent(out))
-    ! function_index=7
+    ! function_index=8
     subroutine sum(values, result)
         use iso_c_binding, only : C_INT, C_SIZE_T
         integer(C_SIZE_T) :: len
@@ -688,7 +751,7 @@ contains
     end subroutine sum
 
     ! bool Function3(bool arg +intent(in)+value)
-    ! function_index=9
+    ! function_index=10
     function function3(arg) &
             result(SHT_rv)
         use iso_c_binding, only : C_BOOL
@@ -702,7 +765,7 @@ contains
     end function function3
 
     ! void Function3b(const bool arg1 +intent(in)+value, bool * arg2 +intent(out), bool * arg3 +intent(inout))
-    ! function_index=10
+    ! function_index=11
     subroutine function3b(arg1, arg2, arg3)
         use iso_c_binding, only : C_BOOL
         logical, value, intent(IN) :: arg1
@@ -722,7 +785,7 @@ contains
 
     ! const std::string Function4a(const std::string & arg1 +intent(in), const std::string & arg2 +intent(in)) +len(30)
     ! arg_to_buffer
-    ! function_index=11
+    ! function_index=12
     function function4a(arg1, arg2) &
             result(SHT_rv)
         use iso_c_binding, only : C_CHAR, C_INT
@@ -738,7 +801,7 @@ contains
 
     ! void Function4b(const std::string & arg1 +intent(in)+len_trim(Larg1), const std::string & arg2 +intent(in)+len_trim(Larg2), std::string & output +intent(out)+len(Noutput))
     ! arg_to_buffer - arg_to_buffer
-    ! function_index=49
+    ! function_index=50
     subroutine function4b(arg1, arg2, output)
         use iso_c_binding, only : C_INT
         character(*), intent(IN) :: arg1
@@ -753,7 +816,7 @@ contains
 
     ! double Function5()
     ! has_default_arg
-    ! function_index=37
+    ! function_index=38
     function function5() &
             result(SHT_rv)
         use iso_c_binding, only : C_DOUBLE
@@ -765,7 +828,7 @@ contains
 
     ! double Function5(double arg1=3.1415 +intent(in)+value)
     ! has_default_arg
-    ! function_index=38
+    ! function_index=39
     function function5_arg1(arg1) &
             result(SHT_rv)
         use iso_c_binding, only : C_DOUBLE
@@ -777,7 +840,7 @@ contains
     end function function5_arg1
 
     ! double Function5(double arg1=3.1415 +intent(in)+value, bool arg2=true +intent(in)+value)
-    ! function_index=13
+    ! function_index=14
     function function5_arg1_arg2(arg1, arg2) &
             result(SHT_rv)
         use iso_c_binding, only : C_BOOL, C_DOUBLE
@@ -793,7 +856,7 @@ contains
 
     ! void Function6(const std::string & name +intent(in))
     ! arg_to_buffer
-    ! function_index=14
+    ! function_index=15
     subroutine function6_from_name(name)
         use iso_c_binding, only : C_INT
         character(*), intent(IN) :: name
@@ -804,7 +867,7 @@ contains
     end subroutine function6_from_name
 
     ! void Function6(int indx +intent(in)+value)
-    ! function_index=15
+    ! function_index=16
     subroutine function6_from_index(indx)
         use iso_c_binding, only : C_INT
         integer(C_INT), value, intent(IN) :: indx
@@ -815,7 +878,7 @@ contains
 
     ! void Function7(int arg +intent(in)+value)
     ! cxx_template
-    ! function_index=39
+    ! function_index=40
     subroutine function7_int(arg)
         use iso_c_binding, only : C_INT
         integer(C_INT), value, intent(IN) :: arg
@@ -826,7 +889,7 @@ contains
 
     ! void Function7(double arg +intent(in)+value)
     ! cxx_template
-    ! function_index=40
+    ! function_index=41
     subroutine function7_double(arg)
         use iso_c_binding, only : C_DOUBLE
         real(C_DOUBLE), value, intent(IN) :: arg
@@ -837,7 +900,7 @@ contains
 
     ! int Function8()
     ! cxx_template
-    ! function_index=41
+    ! function_index=42
     function function8_int() &
             result(SHT_rv)
         use iso_c_binding, only : C_INT
@@ -849,7 +912,7 @@ contains
 
     ! double Function8()
     ! cxx_template
-    ! function_index=42
+    ! function_index=43
     function function8_double() &
             result(SHT_rv)
         use iso_c_binding, only : C_DOUBLE
@@ -861,7 +924,7 @@ contains
 
     ! void Function9(float arg +intent(in)+value)
     ! fortran_generic
-    ! function_index=59
+    ! function_index=60
     subroutine function9_float(arg)
         use iso_c_binding, only : C_DOUBLE, C_FLOAT
         real(C_FLOAT), value, intent(IN) :: arg
@@ -872,7 +935,7 @@ contains
 
     ! void Function9(double arg +intent(in)+value)
     ! fortran_generic
-    ! function_index=60
+    ! function_index=61
     subroutine function9_double(arg)
         use iso_c_binding, only : C_DOUBLE
         real(C_DOUBLE), value, intent(IN) :: arg
@@ -882,7 +945,7 @@ contains
     end subroutine function9_double
 
     ! void Function10()
-    ! function_index=19
+    ! function_index=20
     subroutine function10_0()
         ! splicer begin function.function10_0
         call c_function10_0()
@@ -891,7 +954,7 @@ contains
 
     ! void Function10(const std::string & name +intent(in), float arg2 +intent(in)+value)
     ! fortran_generic - arg_to_buffer
-    ! function_index=61
+    ! function_index=62
     subroutine function10_1_float(name, arg2)
         use iso_c_binding, only : C_DOUBLE, C_FLOAT, C_INT
         character(*), intent(IN) :: name
@@ -904,7 +967,7 @@ contains
 
     ! void Function10(const std::string & name +intent(in), double arg2 +intent(in)+value)
     ! fortran_generic - arg_to_buffer
-    ! function_index=62
+    ! function_index=63
     subroutine function10_1_double(name, arg2)
         use iso_c_binding, only : C_DOUBLE, C_INT
         character(*), intent(IN) :: name
@@ -917,7 +980,7 @@ contains
 
     ! int overload1(int num +intent(in)+value)
     ! has_default_arg
-    ! function_index=43
+    ! function_index=44
     function overload1_num(num) &
             result(SHT_rv)
         use iso_c_binding, only : C_INT
@@ -930,7 +993,7 @@ contains
 
     ! int overload1(int num +intent(in)+value, int offset=0 +intent(in)+value)
     ! has_default_arg
-    ! function_index=44
+    ! function_index=45
     function overload1_num_offset(num, offset) &
             result(SHT_rv)
         use iso_c_binding, only : C_INT
@@ -943,7 +1006,7 @@ contains
     end function overload1_num_offset
 
     ! int overload1(int num +intent(in)+value, int offset=0 +intent(in)+value, int stride=1 +intent(in)+value)
-    ! function_index=21
+    ! function_index=22
     function overload1_num_offset_stride(num, offset, stride) &
             result(SHT_rv)
         use iso_c_binding, only : C_INT
@@ -958,7 +1021,7 @@ contains
 
     ! int overload1(double type +intent(in)+value, int num +intent(in)+value)
     ! has_default_arg
-    ! function_index=45
+    ! function_index=46
     function overload1_3(type, num) &
             result(SHT_rv)
         use iso_c_binding, only : C_DOUBLE, C_INT
@@ -972,7 +1035,7 @@ contains
 
     ! int overload1(double type +intent(in)+value, int num +intent(in)+value, int offset=0 +intent(in)+value)
     ! has_default_arg
-    ! function_index=46
+    ! function_index=47
     function overload1_4(type, num, offset) &
             result(SHT_rv)
         use iso_c_binding, only : C_DOUBLE, C_INT
@@ -986,7 +1049,7 @@ contains
     end function overload1_4
 
     ! int overload1(double type +intent(in)+value, int num +intent(in)+value, int offset=0 +intent(in)+value, int stride=1 +intent(in)+value)
-    ! function_index=22
+    ! function_index=23
     function overload1_5(type, num, offset, stride) &
             result(SHT_rv)
         use iso_c_binding, only : C_DOUBLE, C_INT
@@ -1001,7 +1064,7 @@ contains
     end function overload1_5
 
     ! int useclass(const Class1 * arg1 +intent(in)+value)
-    ! function_index=26
+    ! function_index=27
     function useclass(arg1) &
             result(SHT_rv)
         use iso_c_binding, only : C_INT
@@ -1013,7 +1076,7 @@ contains
     end function useclass
 
     ! const Class1 * getclass2()
-    ! function_index=27
+    ! function_index=28
     function getclass2() &
             result(SHT_rv)
         type(class1) :: SHT_rv
@@ -1023,7 +1086,7 @@ contains
     end function getclass2
 
     ! Class1 * getclass3()
-    ! function_index=28
+    ! function_index=29
     function getclass3() &
             result(SHT_rv)
         type(class1) :: SHT_rv
@@ -1034,7 +1097,7 @@ contains
 
     ! int vector_sum(const std::vector<int> & arg +dimension(:)+intent(in))
     ! arg_to_buffer
-    ! function_index=29
+    ! function_index=30
     function vector_sum(arg) &
             result(SHT_rv)
         use iso_c_binding, only : C_INT, C_LONG
@@ -1047,7 +1110,7 @@ contains
 
     ! void vector_iota(std::vector<int> & arg +dimension(:)+intent(out))
     ! arg_to_buffer
-    ! function_index=30
+    ! function_index=31
     subroutine vector_iota(arg)
         use iso_c_binding, only : C_INT, C_LONG
         integer(C_INT), intent(OUT) :: arg(:)
@@ -1058,7 +1121,7 @@ contains
 
     ! void vector_increment(std::vector<int> & arg +dimension(:)+intent(inout))
     ! arg_to_buffer
-    ! function_index=31
+    ! function_index=32
     subroutine vector_increment(arg)
         use iso_c_binding, only : C_INT, C_LONG
         integer(C_INT), intent(INOUT) :: arg(:)
@@ -1069,7 +1132,7 @@ contains
 
     ! int vector_string_count(const std::vector<std::string> & arg +dimension(:)+intent(in))
     ! arg_to_buffer
-    ! function_index=32
+    ! function_index=33
     !>
     !! \brief count number of underscore in vector of strings
     !!
@@ -1087,7 +1150,7 @@ contains
 
     ! void vector_string_fill(std::vector<std::string> & arg +dimension(:)+intent(out))
     ! arg_to_buffer
-    ! function_index=33
+    ! function_index=34
     !>
     !! \brief Fill in arg with some animal names
     !!
@@ -1107,7 +1170,7 @@ contains
 
     ! void vector_string_append(std::vector<std::string> & arg +dimension(:)+intent(inout))
     ! arg_to_buffer
-    ! function_index=34
+    ! function_index=35
     !>
     !! \brief append '-like' to names.
     !!
@@ -1123,7 +1186,7 @@ contains
 
     ! const std::string & LastFunctionCalled() +len(30)
     ! arg_to_buffer
-    ! function_index=36
+    ! function_index=37
     function last_function_called() &
             result(SHT_rv)
         use iso_c_binding, only : C_CHAR, C_INT
@@ -1158,5 +1221,27 @@ contains
             rv = .false.
         endif
     end function class1_ne
+
+    function singleton_eq(a,b) result (rv)
+        use iso_c_binding, only: c_associated
+        type(singleton), intent(IN) ::a,b
+        logical :: rv
+        if (c_associated(a%voidptr, b%voidptr)) then
+            rv = .true.
+        else
+            rv = .false.
+        endif
+    end function singleton_eq
+
+    function singleton_ne(a,b) result (rv)
+        use iso_c_binding, only: c_associated
+        type(singleton), intent(IN) ::a,b
+        logical :: rv
+        if (.not. c_associated(a%voidptr, b%voidptr)) then
+            rv = .true.
+        else
+            rv = .false.
+        endif
+    end function singleton_ne
 
 end module tutorial_mod